+++ trainGameJava/two_Ten.java
@@ -9,24 +9,24 @@ public class two_Ten {
             for(int j = 0; j < 169; j++){          
                 
                 if(a.results[i] + bcd.results[j] == 10){
-                        buffer = String.format("%f + %s\n", a.results[i], bcd.strings.get(j));
+                        buffer = String.format("%d + %s\n",(int) a.results[i], bcd.strings.get(j));
                         s.addString(buffer);
                 }
             
                 if(a.results[i] - bcd.results[j] == 10){
-                        buffer = String.format("%f - %s\n", a.results[i], bcd.strings.get(j));
+                        buffer = String.format("%d - %s\n",(int) a.results[i], bcd.strings.get(j));
                         s.addString(buffer);     
                 }
                 if(a.results[i] * bcd.results[j] == 10){
-                        buffer = String.format("%f * %s\n", a.results[i], bcd.strings.get(j));
+                        buffer = String.format("%d * %s\n",(int) a.results[i], bcd.strings.get(j));
                         s.addString(buffer);
                 }
                 if(a.results[i] / bcd.results[j] == 10){
-                        buffer = String.format("%f / %s\n", a.results[i], bcd.strings.get(j));
+                        buffer = String.format("%d / %s\n",(int) a.results[i], bcd.strings.get(j));
                         s.addString(buffer);
                 }
                 if(Math.pow((a.results[i]),(bcd.results[j])) == 10){
-                        buffer = String.format("%f ^ %s\n", a.results[i], bcd.strings.get(j));
+                        buffer = String.format("%d ^ %s\n",(int) a.results[i], bcd.strings.get(j));
                         s.addString(buffer);
                 }
             }
